Modify hitbox and vector2

+++ src/main/java/it/unibo/geometrybash/model/geometry/Vector2.java
@@ -10,4 +10,4 @@ package it.unibo.geometrybash.model.geometry;
  * @param x the X coordinate
  * @param y the Y coordinate
  */
-public record Vector2(int x, int y) { }
+public record Vector2(float x, float y) { }
